Add files via upload
how to know the source of the contigs, we use two methods: align contigs to ref transcripts, align reads to contigs. The results from the two methods most are diff. Use this script to statitics the number of contigs which are same source and the number of diff. And statistic the Iden(align to transcripts) and reliablity (align to contigs) in same and diff contigs.